@@ -8,21 +8,25 @@ namespace DeviceApiTester.Infrastructure
 {
     public static class DriverFactory
     {
-        public static InstanceType CreateFromEnum<InstanceType, EnumType>(EnumType driver, params object[] parameters)
+        public static InstanceType? CreateFromEnum<InstanceType, EnumType>(EnumType driver, params object[] parameters)
             where InstanceType : class
         {
             try
             {
+                string name = driver?.ToString() ?? "Foo";
                 ImplementationTypeAttribute creatorAttribute = typeof(EnumType)
-                    .GetMember(driver.ToString())?[0]
+                    .GetMember(name)[0]
                     .GetCustomAttributes(typeof(ImplementationTypeAttribute), false)
                     .OfType<ImplementationTypeAttribute>()
                     .FirstOrDefault()
                     ?? throw new InvalidOperationException($"The {typeof(EnumType).Name}.{driver} enum value is not attributed with an {nameof(ImplementationTypeAttribute)}.");
 
-                return creatorAttribute.ImplementationType == null
-                    ? null
-                    : (InstanceType)Activator.CreateInstance(creatorAttribute.ImplementationType, parameters);
+                if (creatorAttribute.ImplementationType is null)
+                {
+                    return null;
+                }
+
+                return Activator.CreateInstance(creatorAttribute.ImplementationType, parameters) as InstanceType;
             }
             catch (Exception ex)
             {
