@@ -23,6 +23,11 @@ namespace DeviceApiTester.Commands.Script
         {
             Console.WriteLine($"File={ScriptFilePath}");
 
+            if (ScriptFilePath is null)
+            {
+                throw new Exception($"{nameof(ScriptFilePath)} is null");
+            }
+
             try
             {
                 string scriptContents = File.ReadAllText(ScriptFilePath);
@@ -40,6 +45,6 @@ namespace DeviceApiTester.Commands.Script
         /// The file path of script to execute.
         /// </summary>
         [Option('f', "file-path", HelpText = "The file path of script to execute.", Required = true)]
-        public string ScriptFilePath { get; set; }
+        public string? ScriptFilePath { get; set; }
     }
 }
