@@ -12,13 +12,18 @@ namespace DeviceApiTester.Commands.Spi
     public class SpiWriteBytes : SpiCommand, ICommandVerb
     {
         [Option('h', "hex-string", HelpText = "The hexadecimal string to convert and write.  Each byte in string must be represented by two hexadecimal characters.", Required = true)]
-        public string HexString { get; set; }
+        public string? HexString { get; set; }
 
         /// <summary>Executes the command.</summary>
         /// <returns>The command's exit code.</returns>
         public int Execute()
         {
             Console.WriteLine($"BusId={BusId}, ChipSelectLine={ChipSelectLine}, Mode={Mode}, DataBitLength={DataBitLength}, ClockFrequency={ClockFrequency}, HexString={HexString}");
+
+            if (HexString is null)
+            {
+                throw new Exception($"{nameof(HexString)} is null.");
+            }
 
             var connectionSettings = new SpiConnectionSettings(BusId, ChipSelectLine)
             {
@@ -27,12 +32,10 @@ namespace DeviceApiTester.Commands.Spi
                 Mode = Mode,
             };
 
-            using (var spiDevice = SpiDevice.Create(connectionSettings))
-            {
-                // This will verify value as in hexadecimal.
-                var writeBuffer = HexStringUtilities.HexStringToByteArray(HexString);
-                spiDevice.Write(writeBuffer.AsSpan());
-            }
+            using SpiDevice spiDevice = SpiDevice.Create(connectionSettings);
+            // This will verify value as in hexadecimal.
+            var writeBuffer = HexStringUtilities.HexStringToByteArray(HexString);
+            spiDevice.Write(writeBuffer.AsSpan());
 
             return 0;
         }
