@@ -8,11 +8,11 @@ namespace DeviceApiTester.Infrastructure
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
     public class ImplementationTypeAttribute : Attribute
     {
-        public ImplementationTypeAttribute(Type implementationType)
+        public ImplementationTypeAttribute(Type? implementationType)
         {
             ImplementationType = implementationType;
         }
 
-        public Type ImplementationType { get; }
+        public Type? ImplementationType { get; }
     }
 }
